python/legacy/services/zk-refdata-loader/pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "zk-refdata-loader"
version = "0.1.0"
description = "Reference data loaders for zkbot trading venues."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "zk-proto-betterproto>=0.1.0",
    "pymongo>=4.6",
    "firefly_exchange_client>=0.4",
    "v20>=3.0",
    "httpx>=0.27",
    "loguru>=0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
zk-refdata-loader = "zk_refdata_loader.loader:main"

[tool.hatch.build]
include = ["src/zk_refdata_loader/**", "src/tq_refdata_loader/**"]

[tool.hatch.build.targets.wheel]
packages = ["src/zk_refdata_loader", "src/tq_refdata_loader"]
